refactor: Created `MostViewedFooterLayout`
